+++ CHANGELOG.md
@@ -5,9 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [0.3.0] - 2021-06-03
+### Added
+- Add `skip_outdated()` method to ScheduleIter to override default behaviour in `next()`. By default `next()` will never return a date that is in the past, but instead resume the iteration from the current local time.
 ### Changed
 - Rename Timetable to Schedule
-- Calling next() on ScheduleIter will not set the current iteration base to now() if now() is greater than the base, so stopping the iteration for some time and then resuming may return an OffsetDateTime from the past
 
 ## [0.2.2] - 2021-05-25
 ### Changed